Bugfix
git-svn-id: https://svn.aros.org/svn/aros/trunk@3093 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/intuition/doubleclick.c
@@ -62,7 +62,7 @@
 	sMicros += 1000000 * (sSeconds - base);
 	cMicros += 1000000 * (cSeconds - base);
 
-	base = ABS(sMicros - cMicros);
+	base = ABS((LONG)(sMicros - cMicros));
 
 	return (base <= 500000);
     }
